Fix some GC issues (need to hide lxArgv elements for some reason)
Found bugs in GC by calling scripts that allocate a lot of objects
in a loop (see examples/gc_stress files).

Also, start work on generational collector with separate mark and
sweep phases. This isn't done yet but the beginnings of it are
there.

Also, add --profile-GC flag to see how much time the GC ran for, and
how many times it ran.