Deduplicate string objects, by mallocing them instead of having them in object heap
This saves GC time, as only 1 object gets created per string. It
also saves memory fragmentation in the object heap, but is a bit
slower in non-GC because of lack of arena (heap) allocation for the
ObjStrings, they are malloced now.

To get some time savings, we share the string if we can.