add Object->pushedToStack field for generational GC
In gen GC, we need to take care of per-thread C stack objects that
weren't yet pushed to VM stack. We were graying all these objects,
but it was unnecessary and could be very costly. Now, only values
not yet pushed to VM stack are grayed, so the rest can be collected.