Move local get/set to frame->localsTable
Codegen still needs a lot of work to get this working, and I have to
figure out a good way to get this working for blocks like foreach (k
in a) or for loops. It would be ideal to have a scope associated
with a var table, or maybe just 1 var table per function like
currently but loops like this should call SET_LOCAL internally. I
would then need to figure out how variable shadowing works, if it
would at all.